Refactor: Ubuntu Linux Build
**OVERVIEW**

This branch is a work in progress (WIP) to get Dinodeck building on
Ubuntu Linux. It makes the following changes towards this goal:

* Add a Makefile section for Linux.
* Add `src/make_linux.sh`
* Create a file, `LINUX_BUILD.md` describing pre-requisites for Ubuntu
  build.
* Fix various minor capitalization issues as Linux is incredibly
  sensitive to capitalization of file names.
* Miscellaneous minor changes to get the building working.

**TESTING**

I've managed to get Dinodeck building on Ubuntu Linux locally. Sadly, it
usually crashes with a core dump shortly after startup. This appears to
be due to some issues with SDL v1.2 and GLX - namely, it doesn't seem to
recognize that it has a hardware accelerated video card it can use. I'm
still investigating, but for now I haven't personally been able to
run the program.

That being said, this is hopefully a start.